requirements.txt
discord.py==2.2.2
Pillow==10.0.1
pytest==7.4.2
python-dotenv==0.20.0
requests==2.32.3
